feat!: web socket client
The client connects with the web socket server for syncing. An alternative
will be added that uses regular HTTP requests. As part of this work
some missed details came to light that needed addressing.

In order to properly evaluate each incoming change, it's necessary to know
the value of the vclock at the time of the change. I knew the implementation
was necessary but somehow it got lost among all the other details. This
change adds that missing piece, updating `*_changes` and `*_pending` tables
to simplify some of the querying.